Add record type names to filter list
We're not using `business_process_exists` in any meaningful way just
yet, but I'm leaving it in place in case there's some pushback to
letting deploy_opp_record_types fail noisily when it's missing the NPSP
business process. I also wonder whether we should refactor that to use
a GEM business process when `NPSP_Default` is not present.